Fixing branch name
Please pay attention that the URL for the raw file is based on the
branch. This commit changes the README to have `master` as the branch
assuming that it will be merged to it. You will need to put the right
branch name when trying to remote executing the installer from any other
branch.